AddonLoader: Support most addons unaware of delayed loading quirks by emulating PLAYER_LOGIN, PLAYER_ENTERING_WORLD and other events expected by addons.
Rewrite of condition management
X-LoadOn-*-If: for lua preconditions.
X-LoadOn-*-Init: for initializing the addon if extra action necessary.

Delayed loading is the default, target is to do it after the loading screen.
X-LoadOn: Login  or  X-LoadOn-Login  is to be used instead of  X-LoadOn-Always: delayed
Load after an addon conditionally:
X-LoadOn-Addon: <addonName>
X-LoadOn-Addon-If: <lua condition>

Rewrite of Slash and LDB-Launcher

Public API:
AddonLoader.QueueLoadAddOn(addonName, loadCondition)
AddonLoader.LoadAddOn(addonName, loadCondition)